add resolvers

+++ apps/api/src/models/cooks/cooks.resolver.ts
@@ -1,19 +1,43 @@
-import { Resolver, Query, Mutation, Args } from '@nestjs/graphql'
+import {
+  Resolver,
+  Query,
+  Mutation,
+  Args,
+  ResolveField,
+  Parent,
+} from '@nestjs/graphql'
 import { CooksService } from './cooks.service'
 import { Cook } from './entities/cook.entity'
 import { FindManyCookArgs, FindUniqueCookArgs } from './dto/find.args'
 import { CreateCookInput } from './dto/create-cook.input'
 import { UpdateCookInput } from './dto/update-cook.input'
+import { Kitchen } from '../kitchens/entities/kitchen.entity'
+import {
+  AllowAuthenticated,
+  GetUser,
+} from 'src/common/decorators/auth/auth.decorator'
+import { GetUserType } from '@common-kitchen-org/types'
+import { checkRowLevelPermission } from 'src/common/guards'
+import { PrismaService } from 'src/common/prisma/prisma.service'
 
 @Resolver(() => Cook)
 export class CooksResolver {
-  constructor(private readonly cooksService: CooksService) {}
+  constructor(
+    private readonly cooksService: CooksService,
+    private readonly prisma: PrismaService,
+  ) {}
 
+  @AllowAuthenticated()
   @Mutation(() => Cook)
-  createCook(@Args('createCookInput') args: CreateCookInput) {
+  createCook(
+    @Args('createCookInput') args: CreateCookInput,
+    @GetUser() user: GetUserType,
+  ) {
+    checkRowLevelPermission(user, args.uid)
     return this.cooksService.create(args)
   }
 
+  @AllowAuthenticated('admin')
   @Query(() => [Cook], { name: 'cooks' })
   findAll(@Args() args: FindManyCookArgs) {
     return this.cooksService.findAll(args)
@@ -24,13 +48,28 @@ export class CooksResolver {
     return this.cooksService.findOne(args)
   }
 
+  @AllowAuthenticated()
   @Mutation(() => Cook)
-  updateCook(@Args('updateCookInput') args: UpdateCookInput) {
+  updateCook(
+    @Args('updateCookInput') args: UpdateCookInput,
+    @GetUser() user: GetUserType,
+  ) {
+    checkRowLevelPermission(user, args.uid)
     return this.cooksService.update(args)
   }
 
+  @AllowAuthenticated()
   @Mutation(() => Cook)
-  removeCook(@Args() args: FindUniqueCookArgs) {
+  removeCook(@Args() args: FindUniqueCookArgs, @GetUser() user: GetUserType) {
+    checkRowLevelPermission(user, args.where.uid)
     return this.cooksService.remove(args)
+  }
+
+  @ResolveField(() => Kitchen)
+  kitchen(@Parent() cook: Cook, @GetUser() user: GetUserType) {
+    checkRowLevelPermission(user, cook.uid)
+    return this.prisma.kitchen.findUnique({
+      where: { cookId: cook.uid },
+    })
   }
 }
